refactor: merge reader to tracker
